+++ lib/occams-record/pluck.rb
@@ -2,17 +2,17 @@ module OccamsRecord
   module Pluck
     private
 
-    def pluck_results(results, cols, model: nil)
-      if cols.size == 1
-        pluck_results_single(results, cols[0].to_s, model: model)
+    def pluck_results(results, model: nil)
+      casters = TypeCaster.generate(results.columns, results.column_types, model: model)
+      if results[0]&.size == 1
+        pluck_results_single(results, casters)
       else
-        pluck_results_multi(results, cols.map(&:to_s), model: model)
+        pluck_results_multi(results, casters)
       end
     end
 
     # returns an array of values
-    def pluck_results_single(results, col, model: nil)
-      casters = TypeCaster.generate(results.columns, results.column_types, model: model)
+    def pluck_results_single(results, casters)
       col = results.columns[0]
       caster = casters[col]
       if caster
@@ -26,8 +26,7 @@ module OccamsRecord
     end
 
     # returns an array of arrays
-    def pluck_results_multi(results, cols, model: nil)
-      casters = TypeCaster.generate(results.columns, results.column_types, model: model)
+    def pluck_results_multi(results, casters)
       results.map { |row|
         row.map { |col, val|
           caster = casters[col]