+++ test/pluck_test.rb
@@ -40,27 +40,22 @@ class PluckTest < Minitest::Test
   end
 
   def test_raw_pluck_one
-    names = OccamsRecord.sql("SELECT name FROM widgets ORDER BY name", {}).pluck(:name)
-    assert_equal Widget.order(:name).pluck(:name), names
-  end
-
-  def test_raw_pluck_one_str
-    names = OccamsRecord.sql("SELECT name FROM widgets ORDER BY name", {}).pluck("name")
+    names = OccamsRecord.sql("SELECT name FROM widgets ORDER BY name", {}).pluck
     assert_equal Widget.order(:name).pluck(:name), names
   end
 
   def test_raw_pluck_one_str_with_func
-    name_lengths = OccamsRecord.sql("SELECT LENGTH(name) FROM widgets ORDER BY name", {}).pluck("LENGTH(name)")
+    name_lengths = OccamsRecord.sql("SELECT LENGTH(name) FROM widgets ORDER BY name", {}).pluck
     assert_equal [8, 8, 8, 8, 8, 8, 8], name_lengths
   end
 
   def test_raw_pluck_one_str_with_func_as
-    name_lengths = OccamsRecord.sql("SELECT LENGTH(name) AS len FROM widgets ORDER BY name", {}).pluck(:len)
+    name_lengths = OccamsRecord.sql("SELECT LENGTH(name) AS len FROM widgets ORDER BY name", {}).pluck
     assert_equal [8, 8, 8, 8, 8, 8, 8], name_lengths
   end
 
   def test_raw_multi_mixed
-    results = OccamsRecord.sql("SELECT name, LENGTH(name) FROM widgets ORDER BY name", {}).pluck(:name, "LENGTH(name)")
+    results = OccamsRecord.sql("SELECT name, LENGTH(name) FROM widgets ORDER BY name", {}).pluck
     assert_equal [
       ["Widget A", 8],
       ["Widget B", 8],